combining uccl_ep with pybind_proxy
