+++ ep/bench/test_internode_simple.py
@@ -22,11 +22,10 @@ import sys
 # import deep_ep as ep
 try:
     from uccl import ep
-    from uccl import uccl_ep as ep
 except ImportError as exc:
     import sys
 
-    sys.stderr.write("Failed to import uccl_ep\n")
+    sys.stderr.write("Failed to import uccl.ep\n")
     raise
